+++ app/src/main/rs/fftGpuAdapted.rs
@@ -16,10 +16,10 @@ static float read2DArray(int x,int y, float* array,int Length){
     return array[Length*x + y];
 }
 
-float *Angles;
+double *Angles;
 
 void CalculateAngles(float *v_out, uint32_t x) {
-    float PointsDistance = II_PI / F32_SampleLength * (float) x / PRECISION;
+    float PointsDistance = II_PI / F32_SampleLength * x / PRECISION;
 
     uint16_t y = 0;
     float interaction = 0;
@@ -30,7 +30,7 @@ void CalculateAngles(float *v_out, uint32_t x) {
     }
 }
 
-float *Sample;
+short *Sample;
 void fft(float *v_out, uint32_t x) {
     float x_some = 0;
     for (uint32_t y = 0; y < uint32_t_SampleLength; y++) {
